- Changed the return type of Socket::GetInstanceCount() using ssize_t instead of int. - Added a 'debug' option into the configure script.

+++ include/uhttp/net/Socket.h
@@ -20,7 +20,7 @@ class Socket : public SocketImp
 {
 public:
 
-  static int GetInstanceCount();
+  static ssize_t GetInstanceCount();
   
 public:
